Identity service completed baby yahh 😜😁😎

+++ infrastructure/Services/Authentication/JwtSetting.cs
@@ -5,5 +5,5 @@ public class JwtSetting
     public string Issuer { get; set; }
     public string Audience { get; set; }
     public string Key { get; set; }
-    public double ExpireInMonth { get; set; }
+    public int ExpireInMonth { get; set; }
 }
